requirements.txt
# local package
-e .

# external requirements
geopandas
statsmodels
scikit-learn
pyproj
census
requests
lxml
